Player walking animation and position finished

+++ src/graphics_engine/sheet.rs
@@ -35,7 +35,7 @@ impl TileSheet {
         canvas: &mut graphics::Canvas,
         tile: i32, 
         dest: glam::Vec2,
-        scale: i32,
+        scale: f32,
     ) {
         let rect = self.calc_tile_rect(tile);
         canvas.draw(
@@ -43,7 +43,7 @@ impl TileSheet {
             graphics::DrawParam::new()
                 .src(rect)
                 .dest(dest)
-                .scale([scale as f32 * rect.w, scale as f32 * rect.h])
+                .scale([scale * rect.w, scale * rect.h])
         );
     }
